Refactored utility elements to a partial class.
